Added the ability to send HTML messages
There is now an API call that lets you send HTML messages. The shared
code between this new endpoint and the old endpoint was moved into a
private method called raw_send_message to reduce code duplication and to
make it easier to add extra endpoints that send matrix messages.